Add CSE logic to media upload script and cleanup code
