+++ scripts/s3_media_upload
@@ -11,6 +11,8 @@ import psycopg2
 import tqdm
 import yaml
 
+from encryption.client_side_encryption import ClientSideEncryption
+
 # Schema for our sqlite database cache
 SCHEMA = """
     CREATE TABLE IF NOT EXISTS media (
@@ -302,7 +304,7 @@ def run_check_delete(sqlite_conn, base_path):
     print("Updated", len(deleted), "as deleted")
 
 
-def run_upload(s3, bucket, sqlite_conn, base_path, extra_args, should_delete):
+def run_upload(s3, bucket, sqlite_conn, base_path, extra_args, should_delete, cse_master_key):
     """Entry point for upload command
     """
     total = get_not_deleted_count(sqlite_conn)
@@ -332,14 +334,25 @@ def run_upload(s3, bucket, sqlite_conn, base_path, extra_args, should_delete):
         media_uploaded_files = 0
         media_deleted_files = 0
 
+        if cse_master_key:
+            cse_client = ClientSideEncryption('x')
+
         for rel_file_path in local_files:
             local_path = os.path.join(base_path, rel_file_path)
 
             if not check_file_in_s3(s3, bucket, rel_file_path, extra_args):
                 try:
-                    s3.upload_file(
-                        local_path, bucket, rel_file_path, ExtraArgs=extra_args,
-                    )
+                    if cse_master_key:
+                        with cse_client.encrypt as encryptor:
+                            s3.upload_fileobj(
+                                Fileobj=encryptor, 
+                                Bucket=bucket, 
+                                Key=rel_file_path,
+                                ExtraArgs=extra_args)
+                    else:
+                        s3.upload_file(
+                            local_path, bucket, rel_file_path, ExtraArgs=extra_args,
+                        )
                 except Exception as e:
                     print("Failed to upload file %s: %s", local_path, e)
                     continue
@@ -613,6 +626,7 @@ def main():
             args.base_path,
             extra_args,
             should_delete=args.delete,
+            cse_master_key=args.cse_master_key
         )
         return
